refactoring: poll 정보를 context provider를 사용하도록 수정
apollo를 이용하여 쿼리한 poll를 context provider를 이용하여 pollcontainer component에서 사용하도록 수정
* context, useContext hook, context provider 를 파일로 분리
* PollReducer및 context관련 파일일 model/polls로 이동
* poll data query graphql shcheme을 gqlSchemes.js 로 이동
