+++ frontend/guest-app/src/components/TabGroup/TabGroup.js
@@ -5,10 +5,11 @@ import styled from "styled-components";
 import useTabs from "../../hooks/useTabs.js";
 import TabBody from "./TabBody.js";
 import QuestionContainer from "../QuestionContainer/QuestionContainer.js";
-import PollApollo from "../Poll/PollApollo.js";
 import QnATabIcon from "./QnATabIcon.js";
 import PollTabIcon from "./PollTabIcon.js";
 import QuestionsProvider from "../../models/Questions/QuestionProvider.js";
+import PollsProvider from "../../models/Polls/PollsProvider.js";
+import PollContainer from "../Poll/PollContainer.js";
 
 const TabGroupStyle = styled.div`
 	position: fixed;
@@ -35,7 +36,9 @@ function TabGroup({showQnABadge = true, showPollBadge}) {
 				</QuestionsProvider>
 			</TabBody>
 			<TabBody hidden={tabIdx !== POLL_TAB_IDX}>
-				<PollApollo />
+				<PollsProvider>
+					<PollContainer />
+				</PollsProvider>
 			</TabBody>
 		</TabGroupStyle>
 	);
