+++ frontend/guest-app/src/models/Polls/PollsContext.js
@@ -1,0 +1,5 @@
+import {createContext} from "react";
+
+const PollsContext = createContext({});
+
+export default PollsContext;
