+++ frontend/guest-app/src/models/Polls/PollsReducer.js
