+++ frontend/guest-app/src/models/Polls/usePolls.js
@@ -1,0 +1,8 @@
+import {useContext} from "react";
+import PollsContext from "./PollsContext.js";
+
+function usePolls() {
+	return useContext(PollsContext);
+}
+
+export default usePolls;
